Added app_storage test fixture, updated tests to use it
Some test changes were needed to account for the use of a global Storage object. An app_storage fixture was added, returning a function-scoped Storage object. For tests which access storage directly, that fixture was added to their args and references ito current_app.storage were changed to app_storage. For tests that access storage indirectly (for example, via a test journalist or source application), the Storage.get_default() method must be mocked to return the app_storage fixture. This patch has been added in the source_app and journalist_app fixtures, so it will be applied automatically in tests that use them.