Added option to refresh the global InstanceConfig.
If the instance config is updated in the Journalist Interface, the change should
be reflected in both web applications immediately. To allow for this, an
optional refresh argument forcing a database lookup was added, and the source and
journalist apps were updated to use the refresh option before each request.
Subsequent uses of the InstanceConfig within a request should not need to hit the
database.