Get rid of the need to explicitly flush the database session
From what I can tell, nearly all the calls to db.session.flush() have
been introduced so that a newly created object's autoincrement ID can be
referenced in a second object to be created, most often this is creating
a Reply, and then a corresponding SeenReply.

However, if we just pass the Reply object into SeenReply, rather than
the id (which is None if it hasn't been flushed), SQLAlchemy takes care
of it all for us, properly assigning the correct autoincrement ID using
the specified relationship.

The motivation for doing this is that JournalistLoginAttempt in loaddata
was not properly flushing and was being assigned to a
journalist_id=NULL, which will soon be an error as part of #6192. Rather
than stick another flush in to fix that case, we can just get rid of
nearly all of them.

There is still one explicit flush() left in loaddata.py's
record_source_interaction(), but that's a different pattern than these
so I've left it alone for now.