Speed up `make shellcheck` and add podman support
Instead of spawning a new container/shellcheck process for each file we
want to lint, we can pass the entire list of files to it and let it run
over all of them. On my laptop this sped up `make shellcheck` from 26
seconds to 1.5 seconds!

podman support is added using the same environment variable as #6216,
`USE_PODMAN=1`.