adds "skip to main content" link to all pages that inherit from base templates
In accordance with <https://webaim.org/techniques/skipnav/>, the links
are visually hidden until they receive focus, to benefit users
navigating by keyboard whether or not they are using a screen-reader.

The Source Interface's link is looser than that on the Journalist
Interface: it jumps to a MAIN element that includes some navigational
content, and it doesn't appear at all on the landing page.  However, the
point of a skip link is to skip sections *repeated* across a Web
site, and the Source Interface is small enough, and sources'
interactions with it purposeful enough, that I believe this
implementation meets the spirit of WCAG 2.4.1 for SecureDrop's purposes.