Drop unused flask_testing from test-requirements.in
We have not used the flask_testing library since <https://github.com/freedomofpress/securedrop/pull/2948>
(2018), a small part of it was directly copied into SecureDrop.

Plus that version of the library was incompatible with Flask 2.0+ and
by getting rid of it we don't explicitly have to pin Jinja nor markupsafe.