Fix remove redundant (and incorrect) hardcoded time zone hint
The ISO 8601 datetime format can use 'Z' in place of '+00:00',
but that's something that the datetime.isformat() method
will take care of when relevant.

Before this fix, datetimes were output in the (invalid) following format:

    2022-02-09T05:22:12.165773+00:00Z

The ISO 8601 documentation is behind a pay wall.
See https://en.wikipedia.org/wiki/ISO_8601

Co-Authored-By: Allie <allie@freedom.press>