Add honeypot to Source Interface to stop very basic spambots
Add a hidden, empty input field to the Source Interface form that will
abort if anything is submitted in it. This will hopefully stop the most
basic of spambots that fill in spam in every form field they see and
submit the form. It is unlikely to stop any spambot specifically
designed for SecureDrop, as those will adapt as necessary.

This is inspired by and copied from MediaWiki's "SimpleAntiSpam"
functionality originally written by Ryan Schmidt[1].

[1] https://www.mediawiki.org/w/index.php?oldid=797352

Fixes #6295.