Use JavaScript to detect if accessing SI over Tor2Web
For people who are accidentally accessing SecureDrop over Tor2Web (e.g.
by finding it in a search engine), it's probable that they are using a
normal browser and still have JavaScript installed. So we use
`location.href` to detect the URL they accessed the SecureDrop instance
over and redirect to the `/tor2web-warning` endpoint if the URL's
hostname doesn't end in `.onion` (with an exception for localhost).

This is intended to be a best-effort measure, as there's nothing
stopping the Tor2Web proxy from intercepting/modifying our JavaScript.

Fixes #6294.