Use Tor2Web's URL mangling to detect it
Tor2Web mangles onion URLs to point to the proxy, but its implementation
is buggy and we can use that for detection. In this case, it rewrites
a literal `href="fake.onion"` (inside a hidden form <input>) to
something like: `href="//{$address}.onion.ly/fake.onion.ly"`. On form
submission, we can detect that it doesn't roundtrip properly and flag
the mangling as Tor2Web usage.

Since we want to show users the error as soon as possible, we turn the
initial "Get started" button link into a <form> with this special hidden
<input> that POSTs to `/generate`, which will check if Tor2Web mangles
the field and display the warning accordingly.

Fixes #6293.