Prevent viewport jumps when there's flashed messages
Jumping directly to the #flashed anchor on redirect introduced
unintended browser behaviour that pushed the messages we wanted users to
focus on out of view. Since this is an accessibility feature, we don't
want to lose the ability for screen-reader users to jump directly to
what's important immediately, so this introduces a new "Skip to
notification" link, which like the "Skip to main content" link is
visually hidden until in focus.