Make sources.filesystem_id non-nullable
We already set this field all the time and in all places expect it to be
a string, not checking if its None first.

The database migration will delete any sources that have a NULL filesystem_id,
plus anything that references those sources.

Discovered by the mypy work I'm doing in #6346.

Refs #6226.