Fix deduping in test_schema_unchanged_after_up_then_downgrade
The deduping added in fb36b50afbfe96 only applied to the alembic output.
But when we add another schema change on top of it, the test starts
failing because the right side isn't also being de-duped. The easiest
thing to do is dedupe that as well.

While we're at it, improve the test to use normal assertions so that
pytest can print out the exact values that aren't matching, rather than
*only* the schemas, which in this case were actually identical.

Fixes #6352.