Include securedrop-app-code deps into develop-requirements.txt
We want the development virtualenv to include the production dependencies
as well as the development ones for the benefit of tools like mypy as well
as editors like PyCharm. Care was taken to ensure all the versions match
those used in production.

The main change to develop-requirements.txt is that click has been upgraded
to v8, because Flask v2 requires it.

We also have `make venv` downgrade setuptools before we install anything,
this is mostly for the benefit of `jsmin` which uses now-removed setuptools
functionality and requires an older version just to be installable.

Fixes #6358.