SI screen-reader improvements and other smaller fixes
Codename is now not read if "Show Codename" is unchecked, pre-load
images are also not read out at the end of every page.

Also moving codename template to the utils.html, feels more appropriate
if we already have that file anyway

Changes a notification text that refers to the codename "above", but with
these design changes the codename is not above anymore.