Drop Sass
Given what type of web application SecureDrop is, and how the
requirements of the JI and the SI are so disparate, Sass is not
particularly useful to us. Pulling in Dart for that very limited
usefulness seems silly, and LibSass (with its helper tool `sassc`) which
would be available in OS repos is considered deprecated. The status quo
is using the EOL ruby-sass which we definitely do not want, so boring
old CSS it is!