Reorganize Debian packaging, have debhelper do most of the work
Instead of building each package separately and using ansible to
organize files and preprocess templates, use standard Debian
packaging techniques and have debhelper do the heavy lifting.

The packaging is now organized under `securedrop/debian/` and
is all built from one source package. In the future it should be
possible to use standard Debian build tools to build these
packages instead of needing ansible. I also hope that this makes
the actual build process easier to understand, for example, it's
clearer now that we're actually building all of the Python dependencies
from source *twice* (something to fix in a follow-up).

All jinja2 templates have been filled in (all the variables were
fixed/not changing anyways). Control files have been merged into
a single file, the version is now properly read out of the changelog.

Files are installed using `<package>.install`. For convenience, each
package has a folder that matches the hierarchy of how things should
be installed.

The remaining logic still in ansible is compiling of translations, which
is migrated in the next commit for ease of review.

I did not adjust the packaging for the securedrop-grsec metapackage
since it's already being moved into the kernel-builder repository.
Same with the ossec-agent and ossec-server packages, since those are
upstream software they should probably end up in their own repository.

All of the packages except securedrop-app-code are now bit-for-bit
reproducible thanks to all the magic that debhelper does!