Fix some reproducibility issues in securedrop-app-code
The RECORD files are not reproducible. Instead of trying to fix them
up with some sed invocation, we can just delete them, like we do in
all the SecureDrop Workstation packages.

Then there are some known reproducibility issues[1] with the debug
info in the built C extensions. However, the suggested fix of
fixing the file path didn't work because virtualenv uses its own
random file paths that we can't map. In any case, we can just
strip the debug info to remove that source of variance.

This leaves us with only the build ID varying, which is caused by
the same random build path issue. Fixing that will require us to
have reproducible wheels, which is complicated enough that it will
be attempted in a separate PR.

[1] https://reproducible-builds.org/docs/build-path/