Avoid running Python code as root in postinst
There's no need for us to run this as root, all the files alembic needs
to touch (primarily the SQLite database) are writable by www-data. We
specifically run the backup copying and deletion steps as www-data too
so symlink attacks can't be used to have root clobber other files.

For `mod_wsgi-express module-config`, we ship the generated file in the
package itself. In a future (non-hotfix) release we will generate the
file at build-time instead of hardcoding it.