Switch to argon2id passphrase hashes for journalists
argon2id is the current recommended hashing algorithm, as it is able to
take advantage of both the argon2d and argon2i algorithms. It's the
default in argon2-cffi as well.

argon2-cffi automatically handles verifying argon2i passphrases, this is
verified in the new `test_passphrase_argon2i_migration` test case. We
use the `check_needs_rehash()` function to see if it needs an update,
and rehash if so.

valid_password() has been slight adjusted that we return False as soon
as we know the passphrase is invalid rather than needing to guard the
migration checks against is_valid.

The test_valid_login_calls_argon2 test is dropped because it broke with
the new migration checking and IMO provides little value.

Fixes #6655.