Fix PYTHONPATH for scripts in dev environment
`shredder` and `source_deleter` need to be able to import
`journalist_app`, and should have the securedrop/ folder added to
PYTHONPATH, just like is already done for rqrequeue. (rqworker doesn't
need it since that's not our script, it's installed into the venv.)

This better matches how the systemd services are run in production since
the unit has `PYTHONPATH="/var/www/securedrop:..."` (and also the
sys.path addition).

Refs #6588.