Fix permissions on /var/www/securedrop
/var/www/securedrop is now root-owned but world readable. This ensures
that code can only be added/modified/removed by root, not www-data.

The package already installs files /var/www/securedrop with the correct
permissions, we were just clobbering it in the postinst, which is now
removed. We have an explicit chmod to adjust directory permissions.

We need special handling for config.py since it is installed by ansible
instead of the package. Since it contains secrets, we want it to be
only-readable by root/www-data and no one else.

The custom_logo.png is writable by www-data so administrators can change
the logo from the web interface. In the future this could be moved
somewhere else. The file may not exist yet (fresh installs), so we create
it if necessary.

testinfra checks have been updated with the new permissions and took the
time to de-duplicate the list of app directories in 7 different yaml files.