Use shellcheck-py, update to 0.9.0
shellcheck-py is the same shellcheck binary, just packaged into a Python
wheel so it can be pip installed. This makes it much easier to integrate
into our existing Python tooling, letting us skip the need for a system
package or container.

This lets us update to the latest shellcheck, 0.9.0, because we're no
longer limited by what focal has packaged. Two new failures are
suppressed as a result.

While we're at it, clean up the obsolete "shellcheckclean" make target.

Fixes #6715.