Have isort lint everything, move config out of Makefile
Instead of listing the directories to check, have isort check everything
(it automatically excludes folders like .venv). This makes it easier for
people to run isort and ensures we don't forget about any files.

While we're at it, simplify the configuration by setting
profile = "black"[1], which is the main intention of our custom config.

[1] https://pycqa.github.io/isort/docs/configuration/profiles.html