Re-organize ossec packaging to use debhelper and standard layout
Same as the reorganization of the securedrop packages, use a normal
debian/ folder and have debhelper install files into the package.

Hardcode the remaining jinja2 variables and drop the obsolete init.d
scripts.

The systemd units are now installed with dh_installsystemd, so they
end up in the correct /lib/systemd/ rather than /etc/systemd/.

One note is that dh_fixperms is disabled for now just to make the
debdiff readable. It should be re-enabled in a follow-up, soon.