Pull descriptive titles from headings
Per the Web Content Accessibilyt Guidelines 2.0, Success Criterion
2.4.2, all pages should have descriptive titles to help users orient
themselves within the website, particularly for users who take advantage
of screen readers while browsing the web.

The main point of this change is to add a python function that can be
called from jinja2 templates which looks for the first `h1` in the body
block of the post-processed page. That is, if there are 2 `h1`
statements in if/else logic, this function pulls the heading from the
"else" case even though the "if" case appears first in the template.
This functionality requires no special logic in the function, as the
blocks contained within the Context object does not contain text which
is conditionally excluded (this was tested by adding a new user and
selecting a YubiKey secret, then confirming that the page title
correctly stated YubiKey in the confirmation page).

Other changes were needed to support this functionality. Many of the
pages in `securedrop/source_templates` did not have `h1` statements, but
they did have suitable `h2` statements which were trivially converted to
`h1`. However, there are 3 exceptions to this.

`securedrop/source_templates/logout.html` recommends that users refresh
their TOR identity in order to retain anonymity. The heading on the page
contains the text "One more thing...". This soft language seems fine for
fully sighted users because they can quickly skim the page, which
contains very little text. However, a user who is relying on a more
cumbersome screen reader might not be responsive to this relatively soft
language. Therefore, the language has been updated to the stronger
"Logout - Additional Action Needed".

`securedrop/source_templates/tor2web-warning.html` contains the heading
"Proxy Service Detected". The page contains a red warning box drawing
the user's attention to the fact that their connection is not anonymous.
However, a user who cannot see the page would not immediately understand
that there is an important that needs their immediate attention.
Therefore, the language has been updated to say "Warning! Proxy Service
Detected".

Finally, `securedrop/source_templates/index.html` does not contain an h2
block suitable for conversion to an h1 block. Additionally, as this page
does not derive from `base.html`, rather is self-contained, there would
be no body block for the python function to access. Therefore, in this
case, the page title received a hardcoded update adding "Portal" in
place of the contents of an h1 statement.

Fixes: 6313

Signed-off-by: Skyler Ferris <skyvine@protonmail.com>