Use plain container for package building, not molecule
Simplify the package building process by using plain containers that
directly execute `dpkg-buildpackage` with minimal layers of indirection.
Some of the $OCI_BIN, etc. logic is copied or inspired from the
container scripts in securedrop-updater.

The container is explicitly not pinned, instead it's built locally and
if any packages are out of date, rebuilt, before any package building
begins. It's mostly the same, the main difference is that we warm the
cargo cache so it's not fetched cold on each build.

From a developer perspective, it's now trivial to see the build output,
which incidentally reveals some warnings we've been ignoring. OSSEC packages
are also no longer built by default and have separate make targets for
when we want to actually update those.

CI will now build the debs on each commit and PR since it's fast enough
now.

Fixes #6706.