Try GitHub Actions by migrating updater-gui-tests job
The syntax is fairly close, so no actual changes were needed to
the commands that are run.

I did switch usage of `apt` to `apt-get` since the former isn't
supposed to be used in scripts.

I intentionally chose to not migrate the branch filter over to
reduce complexity given how fast and light this job is.