Fix app creation in source.wsgi, add regression test
In journalist.wsgi and source.wsgi, we import the relvant `app` variable
so it can be executed by mod_wsgi. In bea3a0a9a this was accidentally
refactored to be under the `if __name__ == "__main__"` block, so it
wasn't available to be imported.

To prevent this from recurring, add a documentation comment making it
clear the variable needs to be globally accessible.

A regression test mostly verifies that the wsgi files can be executed
without errors. This required a slight modification to
`source_app.create_app()` which tries to scan a directory that doesn't
exist when tests are run. If the folder doesn't exist, we can skip that
entire routine since there's no data to obscure.

Fixes #6741.