Upgrade to mypy 1.0.0
The one issue it flagged is that in journalist_app.create_app(), the
`and handler` check is unnecessary because `handler` will always be
truthy based on the known types of `app.error_handler_spec`.

Once that's removed we can refactor the handler lookup under path
conditional since it's not used otherwise.