Switch cronjobs to systemd timers
systemd timers are easier to define and offer more flexibility in
running commands. For example, 9169606decd3c81c would've been a 2 line
change instead of a complex and fragile bash migration.

We clear the old crontabs in the postinst and let debhelper enable
the timers.

Turns out our usage of `dh_systemd_enable` and `dh_systemd_start`
was wrong, because we were calling it multiple times without specifying
explicit unit files, it was adding the same stanza to the postinst/postrm
multiple times. We now invoke it explicitly for each unit we want
enabled/started. Hopefully when we switch to compat 12/dh_installsystemd
we don't need to manually list each one.

Fixes #6705.
Fixes #6748.