Add accessibility sniffs to pageslayout tests
This commit follows the structure of the current pageslayout tests by
collecting info in the `save_screenshot_and_html` function, and saving
the results into a new subdirectory under
`tests/functional/pageslayout/accessibility-info`. Under this there is a
layer of subdirectories for the locale, then a second layer of
subdirectories to organize messages between errors that require human
review and errors that are machine-verified.

Additionally, this adds a new make target `make accessibility-summary`
which will process all of these data files and display a summary. This
is done rather than displaying the full output because there are
currently ~50,000 lines of output stored in these files; if we displayed
only errors then this would cut it down to 2,600 lines which is a
significant reduction, but still a lot for terminal output. Note that
many of the messages are duplicates across pages. For example, the blue
buttons with white text are slightly below the contrast guideline, and
there is a message about this problem on nearly every page.

This commit also adds html_codesniffer as a dependency to the project.
This is installed globally into the docker image because
html_codesniffer itself requires no dependencies. See PR 6745 for
the full discussion.

Co-authored-by: Cory Francis Myers <cfm@panix.com>