Disable Ubuntu Pro's `ua-timer` and `esm-cache` jobs; causing OSSEC alerts
Ubuntu Pro's "uaclient" expects the Linux kernel package to be versioned
in a specific format that we are not currently compliant with. This
error message is triggering spurious OSSEC alerts.

While we do plan to fix our kernel versioning scheme, we really don't need
this Ubuntu Pro stuff, so let's disable it the same way we do with fwupd.

Since ua-timer is started by a timer, simply disabling it is enough.
But for esm-cache, it's invoked by an apt hook, so we need to mask it.

Fixes #6773.