Have unattended-upgrades remove unused dependencies
By setting `Unattended-Upgrade::Remove-Unused-Dependencies "true";`
unattended-upgrades will remove "unused" dependencies, which are packages
that were pulled in as a dependency but no longer have anything depending
on them and, in theory, can be safely removed.

The main motivation in the short term is to have this remove old
kernel packages (#6762), but this is probably good in the long run too,
since it will ensure the package set on older instances is roughly the
same as fresh installs.

It's definitely possible we depend on something that is just an implicit
dependency and it gets auto removed in the future, but such a bug would
have already caused issues for fresh installs and should be caught by
our nightly staging tests.