Adjust how Rust is installed in the builder container
Install Rust and rustup in slightly more straightforward directories and
unconditionally add it to $PATH instead of weirdly in /root/.cargo.

This setup is partially based on <https://github.com/rust-lang/rustup/issues/1085#issuecomment-296604244>,
which is described as unsupported for multiple users but our containers
are single-user.

The plan is to copy this to the development container as well.