Update Rust in builder to 1.69.0
For the Sequoia work, we're planning to use 1.69.0 so we need to update
the builder to match. This is being done separately since it already
affects a production component (the cryptography wheel) and should be
tested independently.

This also lets us get rid of our index-warming hack since we can use the
new fast sparse protocol instead
(<https://blog.rust-lang.org/2023/03/09/Rust-1.68.0.html#cargos-sparse-protocol>).

Refs #6817.