builder: Fix failure after dependencies are out of date
image_prep.sh is supposed to transparently rebuild the container if it's
out of date, but a logic bug caused that to not work properly.

After the rebuild, we need to reset $status to 0 and then re-run the
dependency check, otherwise it will always fail with an exit code of 42.

(cherry picked from commit f35f400fee622488c34a537173110ca13b4666c5)