fix(ossec): ignore NameError on builtin hasattr()
During mod_wsgi teardown, the Python interpreter may[1] in passing raise
meaningless NameErrors on builtins.  After #6783, we see this on redis's
use of hasattr() in its own teardown.  The NameError itself is a
harmless red herring; here we suppress OSSEC's default alert on the
"error" substring so it doesn't unnecessarily alarm administrators.

[1]: https://stackoverflow.com/a/44940341