Store journalist public key out of the GPG keyring
For Sequoia, we need to have the journalist public key accessible
somewhere that isn't inside the GPG keyring. Storing it on disk as
`journalist.pub` is the simplest solution. This new path is added
to AppArmor.

The initial ansible provisioning is straightforward, we just need to
specify an explicit filename instead of using the admin-provided
filename.

For upgrades it's a bit trickier since we need to get the fingerprint
out of config.py and then have GPG export it.