Drop dh-virtualenv from packaging process
dh-virtualenv is responsible for creating the virtualenv, installing
dependencies from requirements.txt, and then munging the virtualenv to
use the path that it will be installed to.

There are a number of issues with it however:
* largely unmaintained in Debian (missed focal)
* incompatible with debhelper compat 12
* cannot use --require-hashes
* hides deprecation warnings and error messages

It is now getting in the way of Rust packaging since we want to install
a wheel into the virtualenv, however it needs to happen before the
munging step, which dh-virtualenv doesn't really support.

Since we already want to get rid of it, let's just do it now.

Fixes #6538.