Properly exclude custom_logo.png from package and fix test
The test to verify custom_logo.png was not included in the package was
using the wrong path and missed that if a custom_logo.png existed in the
local source tree, it would be included in the package (not an issue for
releases since we always use a clean source tree).

Fix the inclusion by deleting any custom logo if it exists.