Replace flake8 with ruff
ruff is a new Python linter implemented in Rust that's unbelievably fast
and will eventually subsume our usage of pylint and isort. It also makes
it super easy to enable the functionality of some flake8 plugins that we
never did but would have benefitted from.

Fixes identified by ruff and removal of more flake8 suppressions will
come in follow-up commits.