Add columns to `source` table to store PGP public and secret keys
Sequoia doesn't force us to use an on-disk keyring, so we can store the
keys in the database instead. Going forward this will greatly simplify
source deletion since there's no external state that needs to be kept
in sync.

This migration is different because if a downgrade needs to happen for
whatever reason, we do not want to delete the generated key pairs, since
doing so would break migrated sources. It should be safe to leave the
columns in place, so we'll simply do nothing on downgrade.