redwood: Support decrypting messages to multiple recipients
The example code that I based our DecryptionHelper impl on assumed the
message was only encrypted to one recipient, so it hardcoded trying to
decrypt the first pkesk ("Public-Key Encrypted Session Key Packets"[1]).

We now iterate over the pkesks, check whether it matches the key
fingerprint we specified, and if so, attempt to unlock the secret key
and decrypt the message. If no key matches then an error will be
returned.

The existing encryption-decryption test is improved to cover these
cases, by encrypting a message for two recpients and then having both of
those successfully decrypt it. And then the test verifies that when a
third key tries to decrypt, it correctly receives an error.

There is an edge-case if `--throw-keyids` was used to generate the message,
but that shouldn't happen given these are messages encrypted by us without
using that.

[1] https://datatracker.ietf.org/doc/html/rfc4880#section-5.1