redwood: Use keys designated for storage encryption
In practice this makes no difference as the keys we and GPG generate are
marked for both transport and storage encryption, but it clearly
expresses the intent of how this will be used and is straightforward to
add.

While we're here, add the missing filter options from encrypt() to
Helper::decrypt(). In practice this shouldn't make a difference but
is good for consistency. Documentation comments have also been added
to make it explicit that they should be the same.