redwood: Encrypt a stream instead of a file
In the initial implementation, I misunderstood how the interaction
between SecureTemporaryFile and pretty_bad_protocol worked.
SecureTemporaryFile isn't really a file, it's more like a buffer that's
backed by on-disk storage. SecureDrop streamed SecureTemporaryFile into
pretty_bad_protocol, which in turn streamed it to stdin to pass to GPG.

There's no native Python type that represents a stream (ignoring
asyncio), so pyO3 doesn't have a built-in conversion. The simplest thing
is to implement a Rust type that wraps a Python object, calls `.read()`
on it, and implements the `io::Read` trait. The separate pyo3-file[1]
crate (which I found after writing this) does basically the same thing.

The accompanying test case is implemented in Python since we need to
access SecureTemporaryFile.

In theory we could stream files directly into Sequoia and drop the
need for SecureTemporaryFile entirely, but that requires more work
and would need adding compression support to our Rust code.

[1] https://github.com/omerbenamram/pyo3-file